document droplet.make methods
