+++ Sources/Vapor/Service/ServiceError.swift
@@ -1,10 +1,38 @@
 public enum ServiceError: Error, Debuggable {
+    case multipleInstances(
+        type: Any.Type
+    )
+    case disambiguationRequired(
+        key: String,
+        available: [String],
+        type: Any.Type
+    )
+    case unknownService(
+        name: String,
+        available: [String],
+        type: Any.Type
+    )
+    case duplicateServiceName(
+        name: String,
+        type: Any.Type
+    )
+    case noneAvailable(type: Any.Type)
     case unknown(Error)
 }
 
 extension ServiceError {
     public var reason: String {
         switch self {
+        case .multipleInstances(let type):
+            return "Multiple instances available for '\(type)'. Unable to disambiguate."
+        case .noneAvailable(let type):
+            return "No services are available for '\(type)'"
+        case .disambiguationRequired(_, _, let type):
+            return "Multiple services available for '\(type)', please disambiguate using config."
+        case .unknownService(let name, _, let type):
+            return "No service named \"\(name)\" was found while making a '\(type)'."
+        case .duplicateServiceName(let name, let type):
+            return "Duplicate service names were resolved for \"\(name)\" while making a '\(type)'"
         case .unknown(let error):
             return "Unknown: \(error)"
         }
@@ -12,6 +40,16 @@ extension ServiceError {
 
     public var identifier: String {
         switch self {
+        case .multipleInstances:
+            return "multipleInstances"
+        case .noneAvailable:
+            return "none"
+        case .disambiguationRequired:
+            return "disambiguationRequired"
+        case .unknownService:
+            return "unknownService"
+        case .duplicateServiceName:
+            return "duplicate"
         case .unknown:
             return "unknown"
         }
@@ -19,6 +57,18 @@ extension ServiceError {
 
     public var possibleCauses: [String] {
         switch self {
+        case .multipleInstances:
+            return []
+        case .noneAvailable:
+            return [
+                "A provider for this service was not properly configured."
+            ]
+        case .disambiguationRequired:
+            return []
+        case .unknownService:
+            return []
+        case .duplicateServiceName:
+            return []
         case .unknown:
             return []
         }
@@ -26,6 +76,25 @@ extension ServiceError {
 
     public var suggestedFixes: [String] {
         switch self {
+        case .multipleInstances:
+            return [
+                "Register instances as service types instead, so they can be disambiguated using config."
+            ]
+        case .noneAvailable(let type):
+            return [
+                "Register a service that conforms to '\(type)' to the Droplet."
+            ]
+        case .disambiguationRequired(let key, let available, _):
+            return [
+                "Specify one of the available services in `droplet.json` at key `\(key)`.",
+                "Use `try config.set(\"droplet.\(key)\", <service>)` with one of the available service names.",
+                "Available services: \(available)"
+            ]
+        case .unknownService(_, let available, _):
+            let string = available.joined(separator: ", ")
+            return ["Try using one of the available types: \(string)"]
+        case .duplicateServiceName:
+            return []
         case .unknown:
             return []
         }
